Ver 1.1

+++ chickpt.py
@@ -1,31 +1,17 @@
-# 引入函數庫
-import requests
+import re
 import os
+import requests
+import configparser
+
 from time import sleep
 from bs4 import BeautifulSoup
 
-# 定義重整時間
-TIME = 3
-
-# 設定使用者代理，雖然在 小雞上工 中並未限制使用者代理
-headers = {
-    'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36'
-}
-
 
 def readData():
-    existsList = []
-    with open("chickpt.txt", 'r') as f:
-        for line in f.readlines():
-            noEnterLine = line.splitlines()[0]
-            existsList.append(noEnterLine)
-    return existsList
-
-# 發送通知，其中 token 需換成自己的
+    return [l.splitlines()[0] for l in open("chickpt.txt", 'r').readlines()]
 
 
-def notifyLine(message):
-    token = ""
+def notifyLine(token, message):
     headers = {
         "Authorization": "Bearer " + token,
         "Content-Type": "application/x-www-form-urlencoded"
@@ -37,30 +23,52 @@ def notifyLine(message):
 
 
 if __name__ == "__main__":
-    os.system('cls')
-    if not os.path.exists("chickpt.txt"):
+    if not os.path.exists("config.ini") and not os.path.exists("chickpt.txt"):
+        config = configparser.ConfigParser()
+        config["Setting"] = {}
+        config["Setting"]['time'] = "1"
+        config["Setting"]['minSalary'] = "0"
+        config["Setting"]['token'] = ""
+        with open("config.ini", 'w') as f:
+            config.write(f)
+
         open("chickpt.txt", 'w')
-    print(f"小雞上工 監視中，目前是以每 {TIME} 秒檢查一次。")
-    while True:
-        try:
-            url = "https://www.chickpt.com.tw/cases"
+
+    config = configparser.ConfigParser()
+    config.read("config.ini")
+
+    url = "https://www.chickpt.com.tw/cases"
+    headers = {
+        'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36'
+    }
+
+    try:
+        os.system("cls")
+        print(f"小雞上工 監視中，目前是以每 {config['Setting']['time']} 秒檢查一次。")
+        while True:
             existsList = readData()
 
             response = requests.get(url, headers=headers)
             soup = BeautifulSoup(response.text, "html.parser")
 
             for item in soup.find("ul", {"id": "job-list"}):
-                if type(item.find("a")) is int:
+                if item.find("a") == -1:
                     continue
                 aTag = item.find("a")
                 if aTag.get('href') not in existsList:
-                    with open(os.path.join(os.path.abspath("."), "chickpt.txt"), 'a') as f:
+                    salary = aTag.find("span", {"class": "salary"}).text
+                    if salary < config['Setting']['minSalary']:
+                        continue
+                    title = aTag.get("title")
+                    contact = aTag.find(
+                        'p', {'class': 'ellipsis-job-company'}).text.strip()
+                    with open("chickpt.txt", 'a') as f:
                         f.write(f"{aTag.get('href')}\n")
                         print(
-                            f"標題： {aTag.get('title')} 價格： {aTag.find('span', {'class': 'salary'}).text} 聯絡人： {aTag.find('p', {'class': 'ellipsis-job-company'}).text.strip()}")
+                            f"標題： {title} 價格： {aTag.find('span', {'class': 'salary'}).text} 聯絡人： {contact}")
                         notifyLine(
-                            f"\n標題： {aTag.get('title')}\n價格： {aTag.find('span', {'class': 'salary'}).text}\n聯絡人： {aTag.find('p', {'class': 'ellipsis-job-company'}).text.strip()}")
-        except Exception as e:
-            notifyLine("發生錯誤，請盡速處理。")
-            print(e)
-        sleep(TIME)
+                            config['Setting']['token'], f"\n標題： {title}\n價格： {aTag.find('span', {'class': 'salary'}).text}\n聯絡人： {contact}")
+            sleep(int(config['Setting']['time']))
+    except Exception as e:
+        print(f"發生錯誤，錯誤資訊{e}")
+        notifyLine(config['Setting']['token'], f"發生錯誤，請檢查程式。")
